Remove terminal module
