+++ sandbox/continuize_and_render_timer.py
@@ -17,15 +17,15 @@ def main(args):
 
     shot = shot.copy()
 
-    with pt.terminal.TimeCode(success_msg="Trajectories simulated in: "):
+    with pt.utils.TimeCode(success_msg="Trajectories simulated in: "):
         pt.simulate(shot, inplace=True)
 
-    with pt.terminal.TimeCode(success_msg="Trajectories continuized in: "):
+    with pt.utils.TimeCode(success_msg="Trajectories continuized in: "):
         pt.continuize(shot, inplace=True)
 
     class TimedModeManager(ModeManager):
         def change_mode(self, *args, **kwargs):
-            with pt.terminal.TimeCode(success_msg="Animation sequence rendered in: "):
+            with pt.utils.TimeCode(success_msg="Animation sequence rendered in: "):
                 super().change_mode(*args, **kwargs)
             sys.exit()
